+++ src/schemathesis/contrib/openapi/__init__.py
@@ -1,9 +1,11 @@
-from . import formats
+from . import formats, fill_missing_examples
 
 
 def install() -> None:
     formats.install()
+    fill_missing_examples.install()
 
 
 def uninstall() -> None:
     formats.uninstall()
+    fill_missing_examples.uninstall()