+++ test/contrib/openapi/test_fill_missing_examples.py
@@ -1,0 +1,9 @@
+import pytest
+
+
+@pytest.mark.operations("success")
+def test_fills_missing_examples(cli, openapi3_schema_url, snapshot_cli):
+    assert (
+        cli.run(openapi3_schema_url, "--hypothesis-phases=explicit", "--contrib-openapi-fill-missing-examples")
+        == snapshot_cli
+    )